+++ src/Cooked.hs
@@ -9,6 +9,7 @@ where
 
 import Cooked.Attack as X
 import Cooked.Currencies as X
+import Cooked.InitialDistribution as X
 import qualified Cooked.Ltl as Ltl
 import Cooked.MockChain as X
 import Cooked.Output as X
@@ -17,5 +18,6 @@ import Cooked.RawUPLC as X
 import Cooked.ShowBS as X
 import Cooked.Skeleton as X
 import Cooked.Tweak as X
+import Cooked.Validators as X
 import Cooked.ValueUtils as X
 import Cooked.Wallet as X